backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── HTTP client (exchange APIs, sentiment, Telegram) ─────────────────────────
httpx==0.27.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic[email]==2.7.1

# ── WebSocket support ─────────────────────────────────────────────────────────
websockets==12.0
python-multipart==0.0.9

# ── Crypto / signing ─────────────────────────────────────────────────────────
cryptography==42.0.5

# ── Numeric (faster indicator math, optional) ─────────────────────────────────
numpy==1.26.4

# ── Caching / Rate limiting ───────────────────────────────────────────────────
redis==5.0.4
slowapi==0.1.9

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
